+++ sdk/devhub/arm-devhub/CHANGELOG.md
@@ -1,15 +1,33 @@
 # Release History
+    
+## 1.0.0-beta.3 (2023-02-28)
+    
+**Features**
 
-## 1.0.0-beta.3 (Unreleased)
+  - Added Interface ArtifactGenerationProperties
+  - Added Interface GeneratePreviewArtifactsOptionalParams
+  - Added Interface GeneratePreviewArtifactsResponse
+  - Added Type Alias AuthorizationStatus
+  - Added Type Alias DockerfileGenerationMode
+  - Added Type Alias GeneratePreviewArtifactsOperationResponse
+  - Added Type Alias GenerationLanguage
+  - Added Type Alias GenerationManifestType
+  - Added Type Alias ManifestGenerationMode
+  - Interface DeveloperHubServiceClientOptionalParams has a new optional parameter managedClusterResource
+  - Interface Workflow has a new optional parameter artifactGenerationProperties
+  - Added Enum KnownAuthorizationStatus
+  - Added Enum KnownDockerfileGenerationMode
+  - Added Enum KnownGenerationLanguage
+  - Added Enum KnownGenerationManifestType
+  - Added Enum KnownManifestGenerationMode
 
-### Features Added
+**Breaking Changes**
 
-### Breaking Changes
-
-### Bugs Fixed
-
-### Other Changes
-
+  - Class DeveloperHubServiceClient has a new signature
+  - Interface WorkflowListByResourceGroupOptionalParams no longer has parameter managedClusterResource
+  - Type of parameter authStatus of interface Workflow is changed from ManifestType to AuthorizationStatus
+    
+    
 ## 1.0.0-beta.2 (2023-01-04)
     
 **Features**